Streamline contract code, as per review
