+++ common/mixnet-contract/src/mixnode.rs
@@ -236,7 +236,7 @@ pub struct MixNodeBond {
     pub block_height: u64,
     pub mix_node: MixNode,
     pub profit_margin_percent: Option<u8>,
-    pub proxy: Option<Addr>
+    pub proxy: Option<Addr>,
 }
 
 impl MixNodeBond {
@@ -257,7 +257,7 @@ impl MixNodeBond {
             block_height,
             mix_node,
             profit_margin_percent,
-            proxy
+            proxy,
         }
     }
 
@@ -514,7 +514,7 @@ mod tests {
             block_height: 100,
             mix_node: mixnode_fixture(),
             profit_margin_percent: Some(10),
-            proxy: None
+            proxy: None,
         };
 
         let mix2 = MixNodeBond {
@@ -525,7 +525,7 @@ mod tests {
             block_height: 120,
             mix_node: mixnode_fixture(),
             profit_margin_percent: Some(10),
-            proxy: None
+            proxy: None,
         };
 
         let mix3 = MixNodeBond {
@@ -536,7 +536,7 @@ mod tests {
             block_height: 120,
             mix_node: mixnode_fixture(),
             profit_margin_percent: Some(10),
-            proxy: None
+            proxy: None,
         };
 
         let mix4 = MixNodeBond {
@@ -547,7 +547,7 @@ mod tests {
             block_height: 120,
             mix_node: mixnode_fixture(),
             profit_margin_percent: Some(10),
-            proxy: None
+            proxy: None,
         };
 
         let mix5 = MixNodeBond {
@@ -558,7 +558,7 @@ mod tests {
             block_height: 120,
             mix_node: mixnode_fixture(),
             profit_margin_percent: Some(10),
-            proxy: None
+            proxy: None,
         };
 
         // summary:
