+++ contracts/mixnet/src/support/tests.rs
@@ -128,7 +128,7 @@ pub mod test_helpers {
                 ..mix_node_fixture()
             },
             None,
-            None
+            None,
         )
     }
