+++ contracts/vesting/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod contract;
 mod errors;
 pub mod messages;
+mod storage;
 mod support;
 mod traits;
 mod vesting;
