+++ contracts/vesting/src/support/tests.rs
@@ -1,22 +1,11 @@
 #[cfg(test)]
 pub mod helpers {
-    use crate::contract::{instantiate, NUM_VESTING_PERIODS, VESTING_PERIOD};
+    use crate::contract::{instantiate, NUM_VESTING_PERIODS};
     use crate::messages::InitMsg;
-    use crate::vesting::populate_vesting_periods;
-    use crate::vesting::PeriodicVestingAccount;
-    use crate::vesting::VestingPeriod;
+    use crate::vesting::{populate_vesting_periods, Account};
     use config::defaults::DENOM;
-    use cosmwasm_std::from_binary;
-    use cosmwasm_std::testing::mock_dependencies;
-    use cosmwasm_std::testing::mock_env;
-    use cosmwasm_std::testing::mock_info;
-    use cosmwasm_std::testing::MockApi;
-    use cosmwasm_std::testing::MockQuerier;
-    use cosmwasm_std::Addr;
-    use cosmwasm_std::Coin;
-    use cosmwasm_std::OwnedDeps;
-    use cosmwasm_std::Uint128;
-    use cosmwasm_std::{Empty, Env, MemoryStorage, Storage};
+    use cosmwasm_std::testing::{mock_dependencies, mock_env, mock_info, MockApi, MockQuerier};
+    use cosmwasm_std::{Addr, Coin, Empty, Env, MemoryStorage, OwnedDeps, Storage, Uint128};
 
     pub fn init_contract() -> OwnedDeps<MemoryStorage, MockApi, MockQuerier<Empty>> {
         let mut deps = mock_dependencies();
@@ -27,11 +16,11 @@ pub mod helpers {
         return deps;
     }
 
-    pub fn vesting_account_fixture(storage: &mut dyn Storage, env: &Env) -> PeriodicVestingAccount {
+    pub fn vesting_account_fixture(storage: &mut dyn Storage, env: &Env) -> Account {
         let start_time = env.block.time;
         let periods = populate_vesting_periods(start_time.seconds(), NUM_VESTING_PERIODS);
 
-        PeriodicVestingAccount::new(
+        Account::new(
             Addr::unchecked("fixture"),
             Coin {
                 amount: Uint128::new(1_000_000_000_000),
