+++ contracts/vesting/src/traits/bonding_account.rs
@@ -1,5 +1,5 @@
 use crate::errors::ContractError;
-use cosmwasm_std::{Coin, Env, Response, Storage, Timestamp};
+use cosmwasm_std::{Coin, Env, Response, Storage};
 use mixnet_contract::MixNode;
 
 pub trait BondingAccount {
@@ -12,13 +12,6 @@ pub trait BondingAccount {
     ) -> Result<Response, ContractError>;
 
     fn try_unbond_mixnode(&self) -> Result<Response, ContractError>;
-
-    fn track_bond(
-        &self,
-        block_time: Timestamp,
-        bond: Coin,
-        storage: &mut dyn Storage,
-    ) -> Result<(), ContractError>;
 
     fn track_unbond(&self, amount: Coin, storage: &mut dyn Storage) -> Result<(), ContractError>;
 }
